reorganize cpp
