+++ R/RcppExports.R
@@ -1,6 +1,58 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+#' Calculate the Fisher information matrix for a single theta value and a set of items, potentially with a mixture of different models
+#'
+#' @param x Numeric. A single theta value.
+#' @param item_parm A matrix of item parameters.
+#' @template calc-params-mini
+calc_info <- function(x, item_parm, ncat, model) {
+    .Call('_TestDesign_calc_info', PACKAGE = 'TestDesign', x, item_parm, ncat, model)
+}
+
+#' Calculate the Fisher information matrix for a vector of theta values and a set of items, potentially with a mixture of different models
+#'
+#' @param x Numeric. A vector of theta values.
+#' @param item_parm A matrix of item parameters.
+#' @template calc-params-mini
+calc_info_matrix <- function(x, item_parm, ncat, model) {
+    .Call('_TestDesign_calc_info_matrix', PACKAGE = 'TestDesign', x, item_parm, ncat, model)
+}
+
+#' Calculate the Fisher information using empirical Bayes
+#'
+#' Calculate the Fisher information using empirical Bayes.
+#'
+#' @param x A numeric vector of MCMC sampled theta values.
+#' @param item_parm A numeric matrix of item parameters.
+#' @template calc-params-mini
+calc_info_EB <- function(x, item_parm, ncat, model) {
+    .Call('_TestDesign_calc_info_EB', PACKAGE = 'TestDesign', x, item_parm, ncat, model)
+}
+
+#' Calculate the Fisher information using full Bayesian
+#'
+#' Calculate the Fisher information using full Bayesian.
+#'
+#' @param x A numeric vector of MCMC sampled theta values.
+#' @param items_list A list of item parameter matrices.
+#' @template calc-params-mini
+#' @param useEAP \code{TRUE} to use the mean of MCMC theta draws.
+calc_info_FB <- function(x, items_list, ncat, model, useEAP = FALSE) {
+    .Call('_TestDesign_calc_info_FB', PACKAGE = 'TestDesign', x, items_list, ncat, model, useEAP)
+}
+
+#' Calculate the mutual information using full Bayesian
+#'
+#' Calculate the mutual information using full Bayesian.
+#'
+#' @param x A numeric vector of MCMC sampled theta values.
+#' @param items_list A list of item parameter matrices.
+#' @template calc-params-mini
+calc_MI_FB <- function(x, items_list, ncat, model) {
+    .Call('_TestDesign_calc_MI_FB', PACKAGE = 'TestDesign', x, items_list, ncat, model)
+}
+
 #' Calculate probability at a single theta (1PL)
 #'
 #' Calculate the probability of correct response at a theta value, under the 1PL model.
@@ -265,56 +317,15 @@ array_info_gr <- function(x, a, b) {
     .Call('_TestDesign_array_info_gr', PACKAGE = 'TestDesign', x, a, b)
 }
 
-#' Calculate the Fisher information matrix for a single theta value and a set of items, potentially with a mixture of different models
+#' Find the segment to which each theta value belongs
 #'
-#' @param x Numeric. A single theta value.
-#' @param item_parm A matrix of item parameters.
-#' @template calc-params-mini
-calc_info <- function(x, item_parm, ncat, model) {
-    .Call('_TestDesign_calc_info', PACKAGE = 'TestDesign', x, item_parm, ncat, model)
-}
-
-#' Calculate the Fisher information matrix for a vector of theta values and a set of items, potentially with a mixture of different models
+#' Find the segment to which each theta value belongs.
 #'
-#' @param x Numeric. A vector of theta values.
-#' @param item_parm A matrix of item parameters.
-#' @template calc-params-mini
-calc_info_matrix <- function(x, item_parm, ncat, model) {
-    .Call('_TestDesign_calc_info_matrix', PACKAGE = 'TestDesign', x, item_parm, ncat, model)
-}
-
-#' Calculate the Fisher information using empirical Bayes
+#' @param x A numeric vector of theta values.
+#' @param segment A numeric vector of segment cuts.
 #'
-#' Calculate the Fisher information using empirical Bayes.
-#'
-#' @param x A numeric vector of MCMC sampled theta values.
-#' @param item_parm A numeric matrix of item parameters.
-#' @template calc-params-mini
-calc_info_EB <- function(x, item_parm, ncat, model) {
-    .Call('_TestDesign_calc_info_EB', PACKAGE = 'TestDesign', x, item_parm, ncat, model)
-}
-
-#' Calculate the Fisher information using full Bayesian
-#'
-#' Calculate the Fisher information using full Bayesian.
-#'
-#' @param x A numeric vector of MCMC sampled theta values.
-#' @param items_list A list of item parameter matrices.
-#' @template calc-params-mini
-#' @param useEAP \code{TRUE} to use the mean of MCMC theta draws.
-calc_info_FB <- function(x, items_list, ncat, model, useEAP = FALSE) {
-    .Call('_TestDesign_calc_info_FB', PACKAGE = 'TestDesign', x, items_list, ncat, model, useEAP)
-}
-
-#' Calculate the mutual information using full Bayesian
-#'
-#' Calculate the mutual information using full Bayesian.
-#'
-#' @param x A numeric vector of MCMC sampled theta values.
-#' @param items_list A list of item parameter matrices.
-#' @template calc-params-mini
-calc_MI_FB <- function(x, items_list, ncat, model) {
-    .Call('_TestDesign_calc_MI_FB', PACKAGE = 'TestDesign', x, items_list, ncat, model)
+find_segment <- function(x, segment) {
+    .Call('_TestDesign_find_segment', PACKAGE = 'TestDesign', x, segment)
 }
 
 #' Calculate a likelihood value of theta
@@ -481,16 +492,5 @@ theta_FB <- function(nx, theta_init, theta_prop, items_list, item_init, resp, nc
 #' @template calc-params
 theta_FB_single <- function(nx, theta_init, theta_prop, item_mcmc, item_init, resp, ncat, model, prior, prior_parm) {
     .Call('_TestDesign_theta_FB_single', PACKAGE = 'TestDesign', nx, theta_init, theta_prop, item_mcmc, item_init, resp, ncat, model, prior, prior_parm)
-}
-
-#' Find the segment to which each theta value belongs
-#'
-#' Find the segment to which each theta value belongs.
-#'
-#' @param x A numeric vector of theta values.
-#' @param segment A numeric vector of segment cuts.
-#'
-find_segment <- function(x, segment) {
-    .Call('_TestDesign_find_segment', PACKAGE = 'TestDesign', x, segment)
 }
